requirements.txt
tensorflow>=2.14
numpy
pandas
pyyaml
scikit-learn
matplotlib
pillow
opencv-python
streamlit
